+++ packages/dom-snapshot/CHANGELOG.md
@@ -4,6 +4,10 @@
 ## Unreleased
 
 
+## 3.6.2 - 2020/7/7
+
+- fix redirected frames
+
 ## 3.6.1 - 2020/7/5
 
 - fix frames that were redirected with JavaScript ([Trello](https://trello.com/c/egprwtNp))