+++ packages/eyes-images/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 
 
+## 4.14.1 - 2020/7/7
+
+- updated to @applitools/eyes-sdk-core@11.3.0 (from 11.0.5)
+
 ## 4.14.0 - 2020/6/9
 
 - updated to @applitools/eyes-sdk-core@11.0.2 (from v10.2.0)