+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - avoid creating a test if no check command was called ([Trello](https://trello.com/c/ZhKO8sqA/404-wdiojs-sdkempty-test-results-with-new-eyes-core))
 
+## 11.3.1 - 2020/7/7
+
+- fix issue with redirected frames in dom-snapshot ([Trello](https://trello.com/c/egprwtNp))
+
+## 11.3.0 - 2020/7/7
+
+- support visual locators
+
 ## 11.2.2 - 2020/7/5
 
 - cache correct viewport size value after fail set viewport size